Update README to account for new tar build output (#2123)
* Update README to account for new tar build output

The tar output location has been changed and is no longer stable. The old docs were not updated at the time of the change and the result is confusing build breakages. Until the original behavior is added back (not clear that it will happen) the README should document the new behavior.

See Issue #2014 for more info (this was closed as WAI), and [GoogleContainerTools PR#7251](https://github.com/GoogleContainerTools/skaffold/pull/7251) for more context.

* Update README.md

add ".tar" to the query, it is needed!